Added parsing for predicted introns and anticodon
tRNAScan-SE's output file lists the location of possible introns detected within the tRNA coding sequence as well as the anticodon. I have added parsing for the intron location and anticodon and added predicted introns as a split exon in the gff output. The anticodon is added as a qualifier within the tRNA annotation. 
I also modified the naming of the individual features to get it closer to https://www.ncbi.nlm.nih.gov/genbank/eukaryotic_genome_submission_annotation/#rRNA
#hacktoberfest :-)